修正(境界): ublksrv一・八を結べ
Chamado nº 22. A nova biblioteca exterior trouxe extensões GNU, cabeçalhos deliberadamente não estritos e larguras diversas na soleira do núcleo.

A presente carta aparta os avisos estrangeiros sem afrouxar o juízo sobre a nossa lavra, e torna explícitas as conversões reclamadas pela versão 1.8.

Feito por Braga Us <sprtj@protonmail.com>

+++ src/servidor_ublk.c
@@ -424,7 +424,8 @@ int abrir_controle_ublk(struct estado_do_servidor_ublk *servidor)
 
     if (servidor == 0 || servidor->configuracao == 0 ||
         servidor->configuracao->quantidade_de_filas > USHRT_MAX ||
-        servidor->configuracao->quantidade_de_filas > UBLK_MAX_NR_QUEUES ||
+        (unsigned int)servidor->configuracao->quantidade_de_filas >
+            UBLK_MAX_NR_QUEUES ||
         servidor->configuracao->profundidade_das_filas > UBLK_MAX_QUEUE_DEPTH ||
         servidor->configuracao->maior_operacao_em_bytes > INT_MAX) {
         return -EINVAL;
